Finished the demo application.
